+++ test/fixtures/next-app/pages/index.js
@@ -1,10 +1,10 @@
-const image = require('../../../../docs/src/images/coffee1.jpg?widths=100&formats=avif,webp&placeholder=blur')
+const image = require('../../../../docs/src/images/coffee1.jpg?widths=100&formats=avif,webp')
 const { Picture } = require('../../../..')
 
 export default function Home() {
   return (
     <main>
-      <Picture src={image} alt='Coffee' priority />
+      <Picture src={image} alt='Coffee' preload />
       <span data-formats={image.formats.join(',')} data-width={image.width} data-height={image.height} />
     </main>
   )